+++ apps/website/content/blocks/nav-bar.mdx
@@ -6,10 +6,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Left Aligned Navbar
 
-<Demo name="navbars/default01">
+<Demo name="navbars/block01">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default01.tsx",
+    "file": "./src/components/demo/examples/navbars/block01.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
     "codeblock": true
 }
 ```
@@ -17,10 +23,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Center Aligned Navbar
 
-<Demo name="navbars/default02">
+<Demo name="navbars/block02">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default02.tsx",
+    "file": "./src/components/demo/examples/navbars/block02.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
     "codeblock": true
 }
 ```
@@ -28,10 +40,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Right Aligned Navbar
 
-<Demo name="navbars/default03">
+<Demo name="navbars/block03">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default03.tsx",
+    "file": "./src/components/demo/examples/navbars/block03.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
     "codeblock": true
 }
 ```
@@ -39,10 +57,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Navbar with Login Button
 
-<Demo name="navbars/default04">
+<Demo name="navbars/block04">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default04.tsx",
+    "file": "./src/components/demo/examples/navbars/block04.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/block.css",
     "codeblock": true
 }
 ```
@@ -50,10 +74,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Logo Centered Navbar
 
-<Demo name="navbars/default05">
+<Demo name="navbars/block05">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default05.tsx",
+    "file": "./src/components/demo/examples/navbars/block05.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/logo.css",
     "codeblock": true
 }
 ```
@@ -61,10 +91,16 @@ previewImageUrl: /images/nav-bar-preview.svg
 
 ### Page Navigation Navbar
 
-<Demo name="navbars/default06">
+<Demo name="navbars/block06">
 ```json doc-gen:file
 {
-    "file": "./src/components/demo/examples/navbars/default06.tsx",
+    "file": "./src/components/demo/examples/navbars/block06.tsx",
+    "codeblock": true
+}
+```
+```json doc-gen:file
+{
+    "file": "./src/components/demo/examples/navbars/logo.css",
     "codeblock": true
 }
 ```
